+++ fast_rx_persistence/lib/fast_rx_persistence.dart
@@ -1,0 +1,4 @@
+library fast_rx_persistence;
+
+export 'src/exceptions.dart';
+export 'src/fast_rx_persistence_base.dart';