+++ fast_rx_prefs/lib/fast_rx_prefs.dart
@@ -1,5 +1,0 @@
-library fast_rx_prefs;
-
-export 'src/exceptions.dart';
-export 'src/fast_rx_prefs_base.dart';
-export 'src/rx_pref.dart';