+++ fast_ui/example/lib/examples/fast_rx_prefs_example.dart
@@ -1,4 +1,4 @@
-import 'package:fast_rx_prefs/fast_rx_prefs.dart';
+import 'package:fast_rx_persistence/fast_rx_persistence.dart';
 import 'package:fast_ui/fast_ui.dart';
 import 'package:flutter/material.dart';
 
@@ -21,7 +21,7 @@ class FastRxPrefsExample extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text('fast_rx_prefs Example'),
+        title: const Text('fast_rx_persistence Example'),
       ),
       body: Center(
         child: ListView(