+++ fast_rx/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.2.6
+- Export rx.dart to allow for custom extensions in other packages
+
 ## 0.2.5
 - Fixed notifications for `RxList[x] = value`
 - Improved tests to make bugs like that less likely to happen in the future